Fix: Change user review, manner serializer
